refactor(client): move client user store to a separate lib
- [X] break down client-store, move client store to a separate lib;
