+++ angular.json
@@ -830,6 +830,30 @@
       },
       "tags": ["scope:client-store-chatbot", "type:data-access"]
     },
+    "client-store-user": {
+      "projectType": "library",
+      "root": "libs/client-store-user",
+      "sourceRoot": "libs/client-store-user/src",
+      "prefix": "app",
+      "architect": {
+        "test": {
+          "builder": "@nrwl/jest:jest",
+          "outputs": ["coverage/libs/client-store-user"],
+          "options": {
+            "jestConfig": "libs/client-store-user/jest.config.js",
+            "passWithNoTests": true
+          }
+        },
+        "lint": {
+          "builder": "@angular-eslint/builder:lint",
+          "options": {
+            "eslintConfig": "libs/client-store-user/.eslintrc.json",
+            "lintFilePatterns": ["libs/client-store-user/**/*.ts"]
+          }
+        }
+      },
+      "tags": ["scope:client-store-user", "type:data-access"]
+    },
     "client-store-websocket": {
       "architect": {
         "lint": {
