+++ libs/client-components/src/lib/navbar/navbar.component.stories.ts
@@ -3,7 +3,9 @@ import { FlexLayoutModule } from '@angular/flex-layout';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { RouterTestingModule } from '@angular/router/testing';
 import { AppClientMaterialModule } from '@app/client-material';
-import { AppChatbotState, AppSidebarState, AppUserState } from '@app/client-store';
+import { AppSidebarState } from '@app/client-store';
+import { AppChatbotState } from '@app/client-store-chatbot';
+import { AppUserState } from '@app/client-store-user';
 import { documentFactory, WEB_CLIENT_APP_ENV, WINDOW, windowFactory } from '@app/client-util';
 import { NgxsModule } from '@ngxs/store';
 import { Args, Story } from '@storybook/angular/types-6-0';
