+++ libs/client-store-user/.eslintrc.json
@@ -1,0 +1,4 @@
+{
+  "extends": ["../../.eslintrc.js", "../../.eslintrc.angular.js"],
+  "ignorePatterns": ["!**/*"]
+}
