+++ libs/client-store-user/README.md
@@ -1,0 +1,7 @@
+# client-store-user
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test client-store-user` to execute the unit tests.
