+++ libs/client-store-user/src/index.ts
@@ -1,0 +1,3 @@
+export * from './lib/user.actions';
+export * from './lib/user.module';
+export * from './lib/user.store';
