+++ libs/client-store-user/src/lib/user.interface.ts
