+++ libs/client-store-user/src/test-setup.ts
@@ -1,0 +1,4 @@
+import 'jest-preset-angular/setup-jest';
+import { setupJestJsdomGlobalMocks } from '@app/client-unit-testing';
+
+setupJestJsdomGlobalMocks();
