+++ libs/client-store/src/index.ts
@@ -3,4 +3,3 @@ export * from './lib/http-api/index';
 export * from './lib/http-progress/index';
 export * from './lib/sidebar/index';
 export * from './lib/theme/index';
-export * from './lib/user/index';
