+++ libs/client-store/src/lib/client-store.module.ts
@@ -1,11 +1,11 @@
 import { NgModule } from '@angular/core';
 import { AppChatbotStoreModule } from '@app/client-store-chatbot';
+import { AppUserStoreModule } from '@app/client-store-user';
 
 import { AppHttpApStoreModule } from './http-api/http-api.module';
 import { AppHttpProgressStoreModule } from './http-progress/http-progress.module';
 import { AppSidebarStoreModule } from './sidebar/sidebar.module';
 import { AppThemeStoreModule } from './theme/theme.module';
-import { AppUserStoreModule } from './user/user.module';
 
 @NgModule({
   imports: [
