--- libs/client-store/src/lib/user/index.ts
@@ -1,3 +1,0 @@
-export * from './user.actions';
-export * from './user.module';
-export * from './user.store';
