+++ tsconfig.base.json
@@ -46,6 +46,7 @@
       "@app/client-sidebar": ["libs/client-sidebar/src/index.ts"],
       "@app/client-store": ["libs/client-store/src/index.ts"],
       "@app/client-store-chatbot": ["libs/client-store-chatbot/src/index.ts"],
+      "@app/client-store-user": ["libs/client-store-user/src/index.ts"],
       "@app/client-store-websocket": ["libs/client-store-websocket/src/index.ts"],
       "@app/client-translate": ["libs/client-translate/src/index.ts"],
       "@app/client-unit-testing": ["libs/client-unit-testing/src/index.ts"],
